refactor(app): apply architecture audit recommendations
