+++ app/blog/[slug]/page.tsx
@@ -1,24 +1,13 @@
-import { ViewTransition } from "react"
-import type { Metadata, Route } from "next"
-import Link from "next/link"
+import { Suspense } from "react"
+import type { Metadata } from "next"
 import { notFound } from "next/navigation"
 
 import { siteConfig } from "@/lib/config"
-import { calculateReadingTime } from "@/lib/reading-time"
 import {
-  createBlogHref,
-  DraftBadge,
-} from "@/features/post/components/blog-index"
-import { SeriesNav } from "@/features/post/components/series-nav"
-import {
-  formatPostDate,
-  getAllPosts,
-  getPostBySlug,
-  getPostTags,
-  getSeriesPosts,
-} from "@/features/post/post-queries"
-import { MDXContent } from "@/components/mdx-components"
-import MdxLayout from "@/components/mdx-layout"
+  PostDetail,
+  PostDetailSkeleton,
+} from "@/features/post/components/post-detail"
+import { getAllPosts, getPostBySlug } from "@/features/post/post-queries"
 
 export const dynamicParams = false
 
@@ -71,75 +60,11 @@ export const generateMetadata = async (
 }
 
 export default function PostPage({ params }: PageProps<"/blog/[slug]">) {
-  return params.then(({ slug }) => {
-    const post = getPostBySlug(slug)
-    if (!post) notFound()
-
-    const readingTime = calculateReadingTime(post.content)
-    const tags = getPostTags(post.tag)
-    const seriesPosts = post.series
-      ? getSeriesPosts(post.series, post.draft)
-      : []
-    const currentSeriesIndex = seriesPosts.findIndex(
-      (seriesPost) => seriesPost.slug === post.slug
-    )
-
-    return (
-      <article className="mx-auto max-w-4xl py-8 md:py-12">
-        <Link
-          href="/blog"
-          className="mb-8 inline-flex min-h-touch items-center rounded-lg text-sm text-gray-600 transition-colors hover:text-purple-700 focus-visible:outline-hidden focus-visible:ring-2 focus-visible:ring-purple-500/70 focus-visible:ring-offset-2 focus-visible:ring-offset-lightGray dark:text-gray-400 dark:hover:text-purple-300 dark:focus-visible:ring-offset-dark md:mb-10"
-        >
-          Back to articles
-        </Link>
-
-        <header className="mb-8 max-w-3xl md:mb-10">
-          <p className="font-mono text-xs text-gray-500 dark:text-gray-400">
-            <time dateTime={post.date}>{formatPostDate(post.date)}</time>
-            {" · "}
-            {readingTime} min read
-            {post.draft && <DraftBadge />}
-          </p>
-          <ViewTransition
-            name={`post-title-${slug}`}
-            share="text-morph"
-            default="none"
-          >
-            <h1 className="mt-4 text-balance font-heading text-[2rem] font-bold leading-tight tracking-tight text-gray-900 dark:text-ctp-text md:text-5xl">
-              {post.title}
-            </h1>
-          </ViewTransition>
-          <p className="mt-4 max-w-prose text-base leading-relaxed text-gray-600 dark:text-gray-400 md:text-lg">
-            {post.summary}
-          </p>
-          <div className="mt-5 flex flex-wrap gap-2">
-            {tags.map((tag) => (
-              <Link
-                key={tag}
-                href={createBlogHref({ tag })}
-                className="rounded-full border border-purple-500/40 px-3 py-1 text-sm text-purple-700 transition-colors hover:border-purple-500 hover:bg-purple-500/10 focus-visible:outline-hidden focus-visible:ring-2 focus-visible:ring-purple-500/70 dark:border-purple-300/40 dark:text-purple-300 dark:hover:border-purple-300 dark:hover:bg-purple-300/10"
-              >
-                {tag}
-              </Link>
-            ))}
-          </div>
-        </header>
-
-        {post.series && currentSeriesIndex !== -1 && seriesPosts.length > 1 && (
-          <SeriesNav
-            series={post.series}
-            current={currentSeriesIndex + 1}
-            parts={seriesPosts.map((seriesPost) => ({
-              title: seriesPost.title,
-              href: `/blog/${seriesPost.slug}` as Route,
-            }))}
-          />
-        )}
-
-        <MdxLayout>
-          <MDXContent source={post.content} />
-        </MdxLayout>
-      </article>
-    )
-  })
+  return (
+    <Suspense fallback={<PostDetailSkeleton />}>
+      {params.then(({ slug }) => (
+        <PostDetail slug={slug} />
+      ))}
+    </Suspense>
+  )
 }
